requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
mysql-connector-python==8.1.0
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.2
Flask-WTF==1.1.1
WTForms==3.0.1